per-HNB GTP-U traffic counters via nft
Add external dependency libnftables.

When an hNodeB registers, set up netfilter rules to count GTP-U packets
(UDP port 2152) to and from that hNodeB's address -- we are assuming
that it is the same address that Iuh is connecting from.

This is a "workaround" to get performance indicators per hNodeB, without
needing a UPF that supports URR.

Related: SYS#6773
Change-Id: I35b7e97fd039e36633dfde1317170527c82f9f68