use json to decode counters from nftables
Switch nft output to json. Add libjansson dependency, and use it to
parse the result. Navigate the json tree to retrieve the counter values.

Parsing with json takes longer than the direct string parsing of nft
language, but it is the more correct and more stable approach: in case
nftables changes its API, json is likely to still be compatible.

Related: SYS#6773
Change-Id: Id4e7fa017c31945388a010d8581715d71482116b