drop config.vty tests from make check
The intention of those tests was to test various osmo-hnbgw startup
scenarios. It seemed the fastest/simplest way to do it at the time, but
the output depends on another library, regularly causing test fallout.

Keep the tests around for reference. They can still be invoked manually
with 'make config-tests'.

Related: OS#6380
Change-Id: I9c1540904b13d25db1c7933a88f6cc4b028fdd2b