Add osmo_ss7 APIs
Prepare OsmoHNBGW to work with newer versions of libosmo-sigtran, where
struct osmo_ss7_instance was made private. Ensure the APIs are always
present even if building with older libosmo-sigtran, so we can apply a
patch that switches over to using these APIs instead of accessing the
struct directly.

Related: OS#6617
Change-Id: I9e53953bc56fc5920eefb0a39dbf1e0b9843e00e