sccp: Drop error logging for expected scenario
The CN is free to release the SCCP conn at any point through an RLSD,
eg. because the end user on the other side decided to do so.
This was spotted in a osmo-hnbgw deployed in production, where the CN is
sometimes not answering the SCCP-CR[RANAP["Activate PDP Context Req"]]
(only answering with CC but never answering the DTAP on the upper layers).
As a result, most probably the peer times out and decides to
unilaterally release the SCCP conn.

Related: SYS#7454
Change-Id: Ia0941574b05256d6e49babe58df59d938cf637a2