kpi: count RAB actions received in unexpected state
This allows figuring out that something may be wrong in the network,
with both sides (HNB vs CN) going out of sync, or HNBGW ending up in a
bad state due to eg. a bug where a RAB is lost.

While at it, fix some formatting issues on related counters.

Related: SYS#7455
Change-Id: If3521ba601b9d7448226f2303ed2ef402d861bef